// src/app/api/contribuyente/models/buscar-por-documento-response.ts
/* tslint:disable */
/* eslint-disable */
export interface BuscarPorDocumentoResponse {
  contador?: number;
  contribuyenteNumero?: number;
  municipalidadId?: number;
  nombreCompleto?: string;
  nombreCompletoConyuge?: string;
  numOperacionId?: number;
  personaId?: number;
  relContribuyenteNumero?: string;
}