// src/app/api/contribuyente/models/contribuyente-persona-request.ts
/* tslint:disable */
/* eslint-disable */
export interface ContribuyentePersonaRequest {
  apeMaterno?: string;
  apePaterno?: string;
  docIdentidadId: number;
  estCivilTipId?: number;
  fallecio?: number;
  fecFallecimiento?: string;
  fecInscripcion: string;
  fecNacimiento?: string;
  fuenteInfoId?: number;
  genero?: string;
  medDeterminaId: number;
  modOficio?: number;
  nombres?: string;
  numDocIdentidad: string;
  numOperacionId?: number;
  personaId?: number;
  razonSocial?: string;
  segContribuyenteId: number;
  tipMedDeterminaId: number;
  tipPersonaId: number;
}